Refactor TimestampServiceTests
