+++ atlasdb-timelock-server/src/main/java/com/palantir/atlasdb/timelock/atomix/AtomixServerImplementation.java
@@ -109,10 +109,7 @@ public class AtomixServerImplementation implements ServerImplementation {
         DistributedLong timestamp = DistributedValues.getTimestampForClient(replica, client);
         Supplier<TimeLockServices> timeLockSupplier = () -> {
             AtomixTimestampService atomixTimestampService = new AtomixTimestampService(timestamp);
-            return TimeLockServices.create(
-                    atomixTimestampService,
-                    LockServiceImpl.create(),
-                    atomixTimestampService);
+            return TimeLockServices.create(atomixTimestampService, LockServiceImpl.create(), atomixTimestampService);
         };
         return InvalidatingLeaderProxy.create(
                 localMember,
