+++ atlasdb-timelock-server/src/main/java/com/palantir/atlasdb/timelock/atomix/AtomixTimestampService.java
@@ -18,13 +18,12 @@ package com.palantir.atlasdb.timelock.atomix;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.Futures;
-import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampRange;
-import com.palantir.timestamp.TimestampService;
+import com.palantir.timestamp.TimestampServiceWithManagement;
 
 import io.atomix.variables.DistributedLong;
 
-public class AtomixTimestampService implements TimestampService, TimestampManagementService {
+public class AtomixTimestampService implements TimestampServiceWithManagement {
     /**
      * Maximum number of timestamps that may be granted at once.
      */
