+++ atlasdb-timelock-server/src/test/java/com/palantir/atlasdb/timelock/TimeLockResourceTest.java
@@ -23,21 +23,19 @@ import org.junit.Test;
 
 import com.google.common.collect.ImmutableMap;
 import com.palantir.lock.LockService;
-import com.palantir.timestamp.TimestampManagementService;
-import com.palantir.timestamp.TimestampService;
+import com.palantir.timestamp.TimestampServiceWithManagement;
 
 public class TimeLockResourceTest {
     private static final String EXISTING_CLIENT = "existing-client";
     private static final String NON_EXISTING_CLIENT = "non-existing-client";
 
     private static final LockService LOCK_SERVICE = mock(LockService.class);
-    private static final TimestampService TIME_SERVICE = mock(TimestampService.class);
-    private static final TimestampManagementService TIMESTAMP_MANAGEMENT_SERVICE =
-            mock(TimestampManagementService.class);
+    private static final TimestampServiceWithManagement TIMESTAMP_SERVICE_WITH_MANAGEMENT =
+            mock(TimestampServiceWithManagement.class);
     private static final TimeLockServices TIME_LOCK_SERVICES = TimeLockServices.create(
-            TIME_SERVICE,
+            TIMESTAMP_SERVICE_WITH_MANAGEMENT,
             LOCK_SERVICE,
-            TIMESTAMP_MANAGEMENT_SERVICE);
+            TIMESTAMP_SERVICE_WITH_MANAGEMENT);
 
     private static final TimeLockResource RESOURCE = new TimeLockResource(
             ImmutableMap.of(EXISTING_CLIENT, TIME_LOCK_SERVICES));
