+++ timestamp-impl/src/main/java/com/palantir/timestamp/PersistentTimestampService.java
@@ -24,7 +24,7 @@ import com.palantir.common.concurrent.PTExecutors;
 import com.palantir.util.DebugLogger;
 
 @ThreadSafe
-public class PersistentTimestampService implements TimestampService, TimestampManagementService {
+public class PersistentTimestampService implements TimestampServiceWithManagement {
     private static final int MAX_REQUEST_RANGE_SIZE = 10 * 1000;
 
     private final ExecutorService executor;
